Move more actions into ActionBuilder
