+++ dlc-wallet/src/main/scala/org/bitcoins/dlc/wallet/util/DLCActionBuilder.scala
@@ -1,6 +1,7 @@
 package org.bitcoins.dlc.wallet.util
 
 import org.bitcoins.core.api.dlc.wallet.db.DLCDb
+import org.bitcoins.crypto.Sha256Digest
 import org.bitcoins.dlc.wallet.models.{
   DLCAcceptDAO,
   DLCAcceptDb,
@@ -86,5 +87,59 @@ case class DLCActionBuilder(
       .sequence(actions)
       .map(_ => ())
     allActions
+  }
+
+  /** Creates the action to delete the given dlc from our database.
+    * This removes references to the dlc in our various tables
+    */
+  def deleteDLCAction(dlcId: Sha256Digest)(implicit
+      ec: ExecutionContext): DBIOAction[Unit, NoStream, Effect.Write] = {
+    val deleteSigA = dlcSigsDAO.deleteByDLCIdAction(dlcId)
+    val deleteRefundSigA = dlcRefundSigDAO.deleteByDLCIdAction(dlcId)
+    val deleteInputSigA = dlcInputsDAO.deleteByDLCIdAction(dlcId)
+    val deleteAcceptA = dlcAcceptDAO.deleteByDLCIdAction(dlcId)
+    val deleteOfferA = dlcOfferDAO.deleteByDLCIdAction(dlcId)
+    val deleteContractDataA = contractDataDAO.deleteByDLCIdAction(dlcId)
+    val deleteAnnouncementDataA = dlcAnnouncementDAO.deleteByDLCIdAction(dlcId)
+    val deleteDlcA = dlcDAO.deleteByDLCIdAction(dlcId)
+
+    val action = for {
+      _ <- deleteSigA
+      _ <- deleteRefundSigA
+      _ <- deleteInputSigA
+      _ <- deleteAcceptA
+      _ <- deleteOfferA
+      _ <- deleteContractDataA
+      _ <- deleteAnnouncementDataA
+      _ <- deleteDlcA
+    } yield ()
+
+    action
+  }
+
+  /** Retrieves a DBIOAction that fetches the global dlc db,
+    * the contract, the offer, and funding inputs
+    */
+  def getDLCOfferDataAction(dlcId: Sha256Digest)(implicit
+      ec: ExecutionContext): DBIOAction[
+    (
+        Option[DLCDb],
+        Option[DLCContractDataDb],
+        Option[DLCOfferDb],
+        Vector[DLCFundingInputDb]),
+    NoStream,
+    Effect.Read] = {
+    val dlcDbAction = dlcDAO.findByDLCIdAction(dlcId)
+    val contractDataAction = contractDataDAO.findByDLCIdAction(dlcId)
+    val dlcOfferAction = dlcOfferDAO.findByDLCIdAction(dlcId)
+    val fundingInputsAction = dlcInputsDAO.findByDLCIdAction(dlcId)
+    val combined = for {
+      dlcDb <- dlcDbAction
+      contractData <- contractDataAction
+      offer <- dlcOfferAction
+      inputs <- fundingInputsAction
+    } yield (dlcDb, contractData, offer, inputs)
+
+    combined
   }
 }
